add dashBoard to App

+++ lib/features/home/presentation/widgets/product_details_body/buyers_product_rate_panel.dart
@@ -758,7 +758,7 @@ class BuyersProductRatePanel extends StatelessWidget {
 
   Widget _buyerReviewSingle(
     String text,
-    int numOfPercent,
+    double numOfPercent,
     BuildContext context,
   ) {
     return SizedBox(
